Module aws_sdk_rust::aws::common

common contains the type, struct, enum and impls that are common accross most requests such as buckets, objects etc. Contains both Library and Client documentation.

Modules

common
common contains the type, struct, enum and impls that are common accross most requests such as buckets, objects etc. Library Documentation

credentials
credentials contains the type, struct, enum and impls that are credentials related. Client Documentation

macros
macros contains the macros defined for the library.

params
params contains the type, struct, enum and impls that related to URI parameters. Parameters for talking to query-based AWS services.

region
region contains the type, struct, enum and impls for Region related functions. AWS Regions and helper functions.

request
request contains the type, struct, enum and impls that are HTTP Request related. Library Documentation

signature
signature contains the type, struct, enum and impls that are signature related (V2 and V4). This is also where SignedRequest lives which is the core for requests in S3Client. Library Documentation

xmlutil
xmlutil contains the type, struct, enum and impls that are common XML parsing and writing. Library Documentation