Module aws_sdk_rust::aws::common::common

common contains the type, struct, enum and impls that are common accross most requests such as buckets, objects etc. Library Documentation

Used for documenting the library for internal use. Not intended for client documentation. common contains the type, struct, enum and impls that are common accross most requests such as buckets, objects etc.

Structs

BodyParser
Parse Body from XML

BodyWriter
Write Body contents to a SignedRequest

CodeParser
Parse Code from XML

CodeWriter
Write Code contents to a SignedRequest

ContinuationTokenParser
Parse ContinuationToken from XML

ContinuationTokenWriter
Write ContinuationToken contents to a SignedRequest

DisplayNameParser
Parse DisplayName from XML

DisplayNameWriter
Write DisplayName contents to a SignedRequest

EmailAddressParser
Parse EmailAddress from XML

EmailAddressWriter
Write EmailAddress contents to a SignedRequest

HostIdParser
Parse HostId from XML

HostIdWriter
Write HostId contents to a SignedRequest

IDParser
Parse ID from XML

IDWriter
Write ID contents to a SignedRequest

IsTruncatedParser
Parse IsTruncated from XML

IsTruncatedWriter
Write IsTruncated contents to a SignedRequest

KeyCountParser
Parse KeyCount from XML

KeyCountWriter
Write KeyCount contents to a SignedRequest

MaxKeysParser
Parse MaxKeys from XML

MaxKeysWriter
Write MaxKeys contents to a SignedRequest

Operation
Allows for duration tracking of operations. You should not track time of this app running but of each operation and then the summation of the durations plus latency etc. The endpoint etc are Strings instead of Url so it easy to Clone.

Owner
Owner

OwnerParser
Parse Owner from XML

OwnerWriter
Write Owner contents to a SignedRequest

QuietParser
Parse Quiet from XML

QuietWriter
Write Quiet contents to a SignedRequest

RequestIdParser
Parse RequestId from XML

RequestIdWriter
Write RequestId contents to a SignedRequest

ResourceParser
Parse Resource from XML

ResourceWriter
Write Resource contents to a SignedRequest

SkipElementParser
Parse SkipElement from XML

StartAfterParser
Parse StartAfter from XML

StartAfterWriter
Write StartAfter contents to a SignedRequest

TypeParser
Parse Type from XML

TypeWriter
Write Type contents to a SignedRequest

URIParser
Parse URI from XML

URIWriter
Write URI contents to a SignedRequest

ValueParser
Parse Value from XML

ValueWriter
Write Value contents to a SignedRequest

Type Definitions

Body

Code

ContinuationToken

DisplayName

EmailAddress

HostId

ID

IsTruncated

KeyCount

MaxKeys

Quiet

RequestId

Resource

SkipElement

StartAfter

Type

URI

Value