Struct aws_sdk_rust::aws::common::common::Operation

pub struct Operation {
    pub request: String,
    pub endpoint: String,
    pub method: String,
    pub success: bool,
    pub code: u16,
    pub payload_size: u64,
    pub start_time: Option<DateTime<UTC>>,
    pub end_time: Option<DateTime<UTC>>,
    pub duration: Option<Duration>,
    pub object: String,
}

Allows for duration tracking of operations. You should not track time of this app running but of each operation and then the summation of the durations plus latency etc. The endpoint etc are Strings instead of Url so it easy to Clone.

Fields
request: String

Time of operation execution Request (endpoint + path)

endpoint: String

Endpoint URL

method: String

GET, PUT, DELETE...

success: bool

If the operation succeeded or not

code: u16

HTTP return code

payload_size: u64

Size of payload

start_time: Option<DateTime<UTC>>

System time of beginning of actual operation (not time spent on condition logic, etc)

end_time: Option<DateTime<UTC>>

System time of end of actual operation (not time spent on condition logic, etc)

duration: Option<Duration>

Duration of operation

object: String

Object Name