Struct aws_sdk_rust::aws::common::credentials::BaseAutoRefreshingProvider

pub struct BaseAutoRefreshingProvider<P, T> {
    pub credentials_provider: P,
    // some fields omitted
}

Wrapper for AwsCredentialsProvider that caches the credentials returned by the wrapped provider. Each time the credentials are accessed, they are checked to see if they have expired, in which case they are retrieved from the wrapped provider again.

Fields
credentials_provider: P