Struct aws_sdk_rust::aws::common::credentials::EnvironmentProvider

pub struct EnvironmentProvider;

Provides AWS credentials from environment variables. If you decide to use environment variables then the first two listed below are required. The third is used for temporary AWS access and not normally used by third party applications.

AWS_ACCESS_KEY_ID - (required - if using environment variables)

AWS_SECRET_ACCESS_KEY - (required - if using environment variables)

AWS_SESSION_TOKEN - (optional - if using environment variables)

Methods

impl EnvironmentProvider

fn new() -> Result<EnvironmentProvider, CredentialsError>