Type Definition aws_sdk_rust::aws::common::credentials::DefaultCredentialsProviderSync

type DefaultCredentialsProviderSync = AutoRefreshingProviderSync<ChainProvider>;

The credentials provider you probably want to use if you do require your AWS services sync. Wraps a ChainProvider in an AutoRefreshingProvider that uses a Mutex to lock credentials in a threadsafe manner.

The underlying ChainProvider checks multiple sources for credentials, and the AutoRefreshingProvider refreshes the credentials automatically when they expire. The Mutex allows this caching to happen in a Sync manner, incurring the overhead of a Mutex when credentials expire and need to be refreshed.

For a !Sync implementation of the same, see DefaultCredentialsProvider

Methods

impl DefaultCredentialsProviderSync

fn new(
parameters_provider: Option<ParametersProvider>
) -> Result<DefaultCredentialsProviderSync, CredentialsError>