Module aws_sdk_rust::aws::common::signature

signature contains the type, struct, enum and impls that are signature related (V2 and V4). This is also where SignedRequest lives which is the core for requests in S3Client. Library Documentation

AWS API request signatures.

Follows AWS Signature 4 algorithm.

If needed, the request will be re-issued to a temporary redirect endpoint. This can happen with newly created S3 buckets not in us-standard/us-east-1.

Structs

SignedRequest
A data structure for all the elements of an HTTP request that are involved in the Amazon Signature Version 4 signing process version - represents the Signature version. The default is 4 but it can also be set to 2 for older environments.

Functions

string_to_sign_v4
Mark string as AWS4-HMAC-SHA256 hashed