Struct aws_sdk_rust::aws::common::signature::SignedRequest

pub struct SignedRequest<'a> {
    // some fields omitted
}

A data structure for all the elements of an HTTP request that are involved in the Amazon Signature Version 4 signing process version - represents the Signature version. The default is 4 but it can also be set to 2 for older environments.

Methods

impl<'a> SignedRequest<'a>

fn new(method: &str, service: &str, region: Region, bucket: &str, path: &str, signature: &Signature) -> SignedRequest<'a>

Default constructor

fn set_content_type(&mut self, content_type: String)

fn set_signature(&mut self, signature: Signature)

fn set_bucket(&mut self, bucket: &str)

fn set_hostname(&mut self, hostname: Option<String>)

fn set_endpoint_prefix(&mut self, endpoint_prefix: String)

fn set_payload(&mut self, payload: Option<&'a [u8]>)

fn bucket(&self) -> &str

fn method(&self) -> &str

fn signature(&self) -> &Signature

fn path(&self) -> &str

fn canonical_uri(&self) -> &str

fn canonical_query_string(&self) -> &str

fn payload(&self) -> Option<&'a [u8]>

fn headers(&'a self) -> &'a BTreeMap<String, Vec<Vec<u8>>>

fn hostname(&self) -> String

fn remove_header(&mut self, key: &str)

fn add_header(&mut self, key: &str, value: &str)

Add a value to the array of headers for the specified key. Headers are kept sorted by key name for use at signing (BTreeMap)

fn update_header(&mut self, key: &str, value: &str)

fn get_header(&mut self, key: &str) -> String

fn add_param<S>(&mut self, key: S, value: S) where S: Into<String>

fn set_params(&mut self, params: Params)

fn set_ssl(&mut self, ssl: bool)

set_ssl(ssl: bool) - Allows you to override the default ssl flag.

fn ssl(&self) -> bool

ssl() - Returns the ssl bool flag

fn sign(&mut self, creds: &AwsCredentials)