Module aws_sdk_rust::aws::common::xmlutil

xmlutil contains the type, struct, enum and impls that are common XML parsing and writing. Library Documentation

Tools for handling XML from AWS with helper functions for testing.

Wraps an XML stack via traits. Also provides a method of supplying an XML stack from a file for testing purposes.

Structs

XmlParseError
generic Error for XML parsing

XmlResponse
Wraps the Hyper Response type for AWS S3. AWS S3 uses XML instead of JSON.

Traits

Next
Move to the next part of the XML stack

Peek
Peek at next items in the XML stack

Functions

characters
return some XML Characters

end_element
consume an EndElement with a specific name or throw an XmlParseError

end_element_skip
consume an EndElement with a specific name or throw an XmlParseError

indent

optional_string_field
parse Some(String) if the next tag has the right name, otherwise None

peek_at_name
get the name of the current element in the stack. throw a parse error if it's not a StartElement

peek_is_end_element
takes a peek to see if the next element is the end_element

pretty_print_xml

skip_tree
skip a tag and all its children

start_element
consume a StartElement with a specific name or throw an XmlParseError

string_field
return a string field with the right name or throw a parse error

Type Definitions

XmlStack
syntactic sugar for the XML event stack we pass around