Struct aws_sdk_rust::aws::errors::aws::AWSError

pub struct AWSError {
    pub code: String,
    pub host_id: String,
    pub message: String,
    pub request_id: String,
    pub resource: String,
    pub missing_header_name: String,
    pub expanded_message: String,
}

AWSError - Default XML error returned from AWS S3.

AWS returns a core error XML structure plus additional elements based on the type of request. The expanded_message captures those additional elements if you need them.

Fields
code: String

code is an alphanumeric value that in some cases can be a number or a short description of the issue with no spaces.

host_id: String

host_id is the base64 encoded AWS HostId of the node that received the request. This can help AWS track down issues for a given host.

message: String

message is a full description of the error for humans.

request_id: String

request_id is the unique ID generated for the given request coming into AWS. This value is generated on every request coming into AWS to aid in tracking down issues.

resource: String

resource identifies the type of AWS resource.

missing_header_name: String

missing_header_name is added to the default AWSError because it's important for trapping errors where header items have not been added correctly such as acl etc.

expanded_message: String

expanded_message contains the details of a given error message from AWS.

Methods

impl AWSError

AWSError is the generic error XML block that AWS returns on all requests that result in some sort of error.

There are basic core items that return on all AWS errors plus elements that are unique to the given request. For example, if the signature is incorrect (V2 or V4) for some reason then the core fields will be filled in and then the expanded_message field will be updated with the remaining elements that can help you track down the real issue.

NB: AWSError is usually embedded as part of S3Error. S3Error.message will contain the general error message for the given request while AWSError shows the exact AWS error.

fn parse_xml<T: Peek + Next>(tag_name: &str, stack: &mut T) -> Result<AWSError, XmlParseError>