Module aws_sdk_rust::aws::s3::bucket

Structs

Bucket

BucketAlreadyExists
The requested bucket name is not available. The bucket namespace is shared by all users of the system. Please select a different name and try again.

BucketAlreadyExistsParser
Parse BucketAlreadyExists from XML

BucketAlreadyExistsWriter
Write BucketAlreadyExists contents to a SignedRequest

BucketLocationConstraintParser
Parse BucketLocationConstraint from XML

BucketLocationConstraintWriter
Write BucketLocationConstraint contents to a SignedRequest

BucketLoggingStatus

BucketNameParser
Parse BucketName from XML

BucketNameWriter
Write BucketName contents to a SignedRequest

BucketParser
Parse Bucket from XML

BucketWriter
Write Bucket contents to a SignedRequest

BucketsParser
Parse Buckets from XML

BucketsWriter
Write Buckets contents to a SignedRequest

CORSRule

CloudFunctionConfiguration

CloudFunctionConfigurationListParser
Parse CloudFunctionConfigurationList from XML

CloudFunctionConfigurationListWriter
Write CloudFunctionConfigurationList contents to a SignedRequest

CreateBucketRequest

DeleteBucketCorsRequest

DeleteBucketPolicyRequest

DeleteBucketReplicationRequest

DeleteBucketRequest

DeleteBucketTaggingRequest

DeleteBucketWebsiteRequest

Destination

GetBucketCorsOutput

GetBucketCorsRequest

GetBucketLoggingOutput

GetBucketLoggingRequest

GetBucketNotificationConfigurationRequest

GetBucketReplicationOutput

GetBucketVersioningOutput

GetBucketVersioningRequest

HeadBucketRequest

LifecycleConfiguration

LoggingEnabled

NotificationConfiguration
Container for specifying the notification configuration of the bucket. If this element is empty, notifications are turned off on the bucket.

NotificationConfigurationParser
Parse NotificationConfiguration from XML

NotificationConfigurationWriter
Write NotificationConfiguration contents to a SignedRequest

PutBucketLifecycleRequest

PutBucketLoggingRequest

PutBucketNotificationRequest

PutBucketPolicyRequest
Put bucket policy request

PutBucketReplicationRequest

PutBucketRequestPaymentRequest

PutBucketVersioningRequest

PutBucketWebsiteRequest

QueueConfigurationDeprecated

ReplicationConfiguration
Container for replication rules. You can add as many as 1,000 rules. Total replication configuration size can be up to 2 MB.

RoutingRule

TopicConfigurationDeprecated

TopicConfigurationListParser
Parse TopicConfigurationList from XML

TopicConfigurationListWriter
Write TopicConfigurationList contents to a SignedRequest

TopicConfigurationParser
Parse TopicConfiguration from XML

TopicConfigurationWriter
Write TopicConfiguration contents to a SignedRequest

VersioningConfiguration

WebsiteConfiguration

Functions

create_bucket_config_xml
Manually writes out bucket configuration (location constraint) in XML.

needs_create_bucket_config
Helper function to determine if a create config is needed.

Type Definitions

BucketLocationConstraint

BucketName

Buckets

CloudFunctionConfigurationList

TopicConfigurationList