Module aws_sdk_rust::aws::s3::endpoint

Client Documentation

The crate Url is required and used to build and extract portions of the Url as needed by the library. See the README.md and/or src/main.rs for an example of how to use the library.

Structs

Endpoint
Endpoint allows you to set a custom endpoint and/or a proxy for a given region and associate this as an endpoint of where S3Client will look for the data.

Enums

Signature
Required to specify which type of API Signature to use. AWS defaults to using V4 by default. However, third party applications often use V2 (AWS will still honor V2).