Module aws_sdk_rust::aws::s3::object

Library Documentation

This file contains everything related to Objects.

Structs

CompleteMultipartUploadOutput

CompleteMultipartUploadOutputParser
Parse CompleteMultipartUploadOutput from XML

CompleteMultipartUploadOutputWriter
Write CompleteMultipartUploadOutput contents to a SignedRequest

CompleteMultipartUploadRequest

CopyObjectOutput

CopyObjectRequest

CopyObjectResult

CreateMultipartUploadOutput

CreateMultipartUploadRequest

Delete

DeleteMarkerEntry

DeleteObjectOutput

DeleteObjectRequest

DeleteObjectsOutput

DeleteObjectsRequest

DeletedObject

ErrorDocument

ExpiresParser
Parse Expires from XML

ExpiresWriter
Write Expires contents to a SignedRequest

GetObjectOutput

GetObjectRequest

GetObjectRequestParser
Parse GetObjectRequest from XML

GetObjectRequestWriter
Write GetObjectRequest contents to a SignedRequest

GetObjectTorrentRequest

HeadObjectOutput

IndexDocument

Initiator

LambdaFunctionConfiguration
Container for specifying the AWS Lambda notification configuration.

ListMultipartUploadsOutput

ListMultipartUploadsOutputParser
Parse ListMultipartUploadsOutput from XML

ListMultipartUploadsOutputWriter
Write ListMultipartUploadsOutput contents to a SignedRequest

ListObjectVersionsOutput

ListObjectVersionsRequest

ListObjectsOutput

ListObjectsRequest

ListVersionsResult

MaxUploadsParser
Parse MaxUploads from XML

MaxUploadsWriter
Write MaxUploads contents to a SignedRequest

MultipartUpload

MultipartUploadListParser
Parse MultipartUploadList from XML

MultipartUploadListWriter
Write MultipartUploadList contents to a SignedRequest

MultipartUploadParser
Parse MultipartUpload from XML

MultipartUploadWriter
Write MultipartUpload contents to a SignedRequest

NextUploadIdMarkerParser
Parse NextUploadIdMarker from XML

NextUploadIdMarkerWriter
Write NextUploadIdMarker contents to a SignedRequest

NoncurrentVersionExpiration
Specifies when noncurrent object versions expire. Upon expiration, Amazon S3 permanently deletes the noncurrent object versions. You set this lifecycle configuration action on a bucket that has versioning enabled (or suspended) to request that Amazon S3 delete noncurrent object versions at a specific period in the object's lifetime.

Object

ObjectIdentifier

Part

PartNumberParser
Parse PartNumber from XML

PartNumberWriter
Write PartNumber contents to a SignedRequest

PartParser
Parse Part from XML

PartWriter
Write Part contents to a SignedRequest

PutObjectOutput

PutObjectOutputParser
Parse PutObjectOutput from XML

PutObjectOutputWriter
Write PutObjectOutput contents to a SignedRequest

PutObjectRequest

QueueConfiguration
Container for specifying an configuration when you want Amazon S3 to publish events to an Amazon Simple Queue Service (Amazon SQS) queue.

Redirect

RedirectAllRequestsTo

ReplicationRule

RestoreObjectOutput

RestoreObjectRequest

RestoreRequest

Tag

TagParser
Parse Tag from XML

TagSetParser
Parse TagSet from XML

TagSetWriter
Write TagSet contents to a SignedRequest

TagWriter
Write Tag contents to a SignedRequest

TopicConfiguration
Container for specifying the configuration when you want Amazon S3 to publish events to an Amazon Simple Notification Service (Amazon SNS) topic.

UploadIdMarkerParser
Parse UploadIdMarker from XML

UploadIdMarkerWriter
Write UploadIdMarker contents to a SignedRequest

UploadPartRequest

Type Definitions

Expires

MaxUploads

MultipartUploadList

NextUploadIdMarker

ObjectVersionList

PartNumber

TagSet

UploadIdMarker