Struct aws_sdk_rust::aws::s3::object::GetObjectOutput

pub struct GetObjectOutput {
    pub last_modified: LastModified,
    pub content_range: ContentRange,
    pub request_charged: RequestCharged,
    pub content_encoding: ContentEncoding,
    pub replication_status: ReplicationStatus,
    pub storage_class: StorageClass,
    pub server_side_encryption: ServerSideEncryption,
    pub ssekms_key_id: SSEKMSKeyId,
    pub content_disposition: ContentDisposition,
    pub metadata: Metadata,
    pub body: Body,
    pub accept_ranges: AcceptRanges,
    pub website_redirect_location: WebsiteRedirectLocation,
    pub expires: Expires,
    pub delete_marker: DeleteMarker,
    pub cache_control: CacheControl,
    pub content_length: ContentLength,
    pub expiration: Expiration,
    pub missing_meta: MissingMeta,
    pub restore: Restore,
    pub sse_customer_algorithm: SSECustomerAlgorithm,
    pub content_type: ContentType,
    pub content_language: ContentLanguage,
    pub version_id: ObjectVersionId,
    pub e_tag: ETag,
    pub sse_customer_key_md5: SSECustomerKeyMD5,
}

Fields
last_modified: LastModified

Last modified date of the object

content_range: ContentRange

The portion of the object returned in the response.

request_charged: RequestCharged content_encoding: ContentEncoding

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

replication_status: ReplicationStatus storage_class: StorageClass server_side_encryption: ServerSideEncryption

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

ssekms_key_id: SSEKMSKeyId

If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

content_disposition: ContentDisposition

Specifies presentational information for the object.

metadata: Metadata

A map of metadata to store with the object in S3.

body: Body

Object data.

accept_ranges: AcceptRanges website_redirect_location: WebsiteRedirectLocation

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

expires: Expires

The date and time at which the object is no longer cacheable.

delete_marker: DeleteMarker

Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

cache_control: CacheControl

Specifies caching behavior along the request/reply chain.

content_length: ContentLength

Size of the body in bytes.

expiration: Expiration

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key value pairs providing object expiration information. The value of the rule-id is URL encoded.

missing_meta: MissingMeta

This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

restore: Restore

Provides information about object restoration operation and expiration time of the restored object copy.

sse_customer_algorithm: SSECustomerAlgorithm

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

content_type: ContentType

A standard MIME type describing the format of the object data.

content_language: ContentLanguage

The language the content is in.

version_id: ObjectVersionId

Version of the object.

e_tag: ETag

An ETag is an opaque identifier assigned by a web server to a specific version of a resource found at a URL

sse_customer_key_md5: SSECustomerKeyMD5

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.