Struct aws_sdk_rust::aws::s3::object::GetObjectRequest

pub struct GetObjectRequest {
    pub response_content_encoding: Option<ResponseContentEncoding>,
    pub response_content_language: Option<ResponseContentLanguage>,
    pub sse_customer_algorithm: Option<SSECustomerAlgorithm>,
    pub response_content_type: Option<ResponseContentType>,
    pub if_unmodified_since: Option<IfUnmodifiedSince>,
    pub version_id: Option<ObjectVersionId>,
    pub request_payer: Option<RequestPayer>,
    pub response_cache_control: Option<ResponseCacheControl>,
    pub sse_customer_key: Option<SSECustomerKey>,
    pub bucket: BucketName,
    pub if_none_match: Option<IfNoneMatch>,
    pub response_content_disposition: Option<ResponseContentDisposition>,
    pub range: Option<Range>,
    pub key: ObjectKey,
    pub if_match: Option<IfMatch>,
    pub response_expires: Option<ResponseExpires>,
    pub if_modified_since: Option<IfModifiedSince>,
    pub sse_customer_key_md5: Option<SSECustomerKeyMD5>,
}

Fields
response_content_encoding: Option<ResponseContentEncoding>

Sets the Content-Encoding header of the response.

response_content_language: Option<ResponseContentLanguage>

Sets the Content-Language header of the response.

sse_customer_algorithm: Option<SSECustomerAlgorithm>

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

response_content_type: Option<ResponseContentType>

Sets the Content-Type header of the response.

if_unmodified_since: Option<IfUnmodifiedSince>

Return the object only if it has not been modified since the specified time, otherwise return a 412 (precondition failed).

version_id: Option<ObjectVersionId>

VersionId used to reference a specific version of the object.

request_payer: Option<RequestPayer> response_cache_control: Option<ResponseCacheControl>

Sets the Cache-Control header of the response.

sse_customer_key: Option<SSECustomerKey>

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side- encryption-customer-algorithm header.

bucket: BucketName if_none_match: Option<IfNoneMatch>

Return the object only if its entity tag (ETag) is different from the one specified, otherwise return a 304 (not modified).

response_content_disposition: Option<ResponseContentDisposition>

Sets the Content-Disposition header of the response

range: Option<Range>

Downloads the specified range bytes of an object. For more information about the HTTP Range header, go to http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35.

key: ObjectKey if_match: Option<IfMatch>

Return the object only if its entity tag (ETag) is the same as the one specified, otherwise return a 412 (precondition failed).

response_expires: Option<ResponseExpires>

Sets the Expires header of the response.

if_modified_since: Option<IfModifiedSince>

Return the object only if it has been modified since the specified time, otherwise return a 304 (not modified).

sse_customer_key_md5: Option<SSECustomerKeyMD5>

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.