Struct aws_sdk_rust::aws::s3::object::MultipartUploadListOutput

pub struct MultipartUploadListOutput {
    pub upload_id_marker: UploadIdMarker,
    pub common_prefixes: CommonPrefixList,
    pub next_key_marker: NextKeyMarker,
    pub bucket: BucketName,
    pub delimiter: Delimiter,
    pub next_upload_id_marker: NextUploadIdMarker,
    pub prefix: Prefix,
    pub uploads: MultipartUploadList,
    pub key_marker: KeyMarker,
    pub max_uploads: MaxUploads,
    pub encoding_type: EncodingType,
    pub is_truncated: IsTruncated,
}

Fields
upload_id_marker: UploadIdMarker

Upload ID after which listing began.

common_prefixes: CommonPrefixList next_key_marker: NextKeyMarker

When a list is truncated, this element specifies the value that should be used for the key-marker request parameter in a subsequent request.

bucket: BucketName

Name of the bucket to which the multipart upload was initiated.

delimiter: Delimiter next_upload_id_marker: NextUploadIdMarker

When a list is truncated, this element specifies the value that should be used for the upload-id-marker request parameter in a subsequent request.

prefix: Prefix

When a prefix is provided in the request, this field contains the specified prefix. The result contains only keys starting with the specified prefix.

uploads: MultipartUploadList key_marker: KeyMarker

The key at or after which the listing began.

max_uploads: MaxUploads

Maximum number of multipart uploads that could have been included in the response.

encoding_type: EncodingType

Encoding type used by Amazon S3 to encode object keys in the response.

is_truncated: IsTruncated

Indicates whether the returned list of multipart uploads is truncated. A value of true indicates that the list was truncated. The list can be truncated if the number of multipart uploads exceeds the limit allowed or specified by max uploads.