Struct aws_sdk_rust::aws::s3::object::MultipartUploadListPartsOutput

pub struct MultipartUploadListPartsOutput {
    pub initiator: Initiator,
    pub bucket: BucketName,
    pub next_part_number_marker: NextPartNumberMarker,
    pub parts: Parts,
    pub upload_id: MultipartUploadId,
    pub storage_class: StorageClass,
    pub key: ObjectKey,
    pub request_charged: RequestCharged,
    pub owner: Owner,
    pub max_parts: MaxParts,
    pub is_truncated: IsTruncated,
    pub part_number_marker: PartNumberMarker,
}

Fields
initiator: Initiator

Identifies who initiated the multipart upload.

bucket: BucketName

Name of the bucket to which the multipart upload was initiated.

next_part_number_marker: NextPartNumberMarker

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

parts: Parts upload_id: MultipartUploadId

Upload ID identifying the multipart upload whose parts are being listed.

storage_class: StorageClass

The class of storage used to store the object.

key: ObjectKey

Object key for which the multipart upload was initiated.

request_charged: RequestCharged owner: Owner max_parts: MaxParts

Maximum number of parts that were allowed in the response.

is_truncated: IsTruncated

Indicates whether the returned list of parts is truncated.

part_number_marker: PartNumberMarker

Part number after which listing begins.