Struct aws_sdk_rust::aws::s3::object::MultipartUploadListRequest

pub struct MultipartUploadListRequest {
    pub upload_id_marker: Option<UploadIdMarker>,
    pub bucket: BucketName,
    pub delimiter: Option<Delimiter>,
    pub prefix: Option<Prefix>,
    pub key_marker: Option<KeyMarker>,
    pub max_uploads: Option<MaxUploads>,
    pub encoding_type: Option<EncodingType>,
}

Fields
upload_id_marker: Option<UploadIdMarker>

Together with key-marker, specifies the multipart upload after which listing should begin. If key-marker is not specified, the upload-id-marker parameter is ignored.

bucket: BucketName delimiter: Option<Delimiter>

Character you use to group keys.

prefix: Option<Prefix>

Lists in-progress uploads only for those keys that begin with the specified prefix.

key_marker: Option<KeyMarker>

Together with upload-id-marker, this parameter specifies the multipart upload after which listing should begin.

max_uploads: Option<MaxUploads>

Sets the maximum number of multipart uploads, from 1 to 1,000, to return in the response body. 1,000 is the maximum number of uploads that can be returned in a response.

encoding_type: Option<EncodingType>