Struct aws_sdk_rust::aws::s3::object::PutObjectOutput

pub struct PutObjectOutput {
    pub sse_customer_algorithm: SSECustomerAlgorithm,
    pub request_charged: RequestCharged,
    pub version_id: ObjectVersionId,
    pub e_tag: ETag,
    pub expiration: Expiration,
    pub server_side_encryption: ServerSideEncryption,
    pub sse_customer_key_md5: SSECustomerKeyMD5,
    pub ssekms_key_id: SSEKMSKeyId,
}

Fields
sse_customer_algorithm: SSECustomerAlgorithm

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

request_charged: RequestCharged version_id: ObjectVersionId

Version of the object.

e_tag: ETag

Entity tag for the uploaded object.

expiration: Expiration

If the object expiration is configured, this will contain the expiration date (expiry-date) and rule ID (rule-id). The value of rule-id is URL encoded.

server_side_encryption: ServerSideEncryption

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

sse_customer_key_md5: SSECustomerKeyMD5

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

ssekms_key_id: SSEKMSKeyId

If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.