Module aws_sdk_rust::aws::s3::writeparse

Library Documentation

This file is a WIP file. It contains a lot of technical debt that is in the process of being cleaned up and organized. It will most likely go away in the near future.

Structs

AllowedMethodParser
Parse AllowedMethod from XML

AllowedMethodWriter
Write AllowedMethod contents to a SignedRequest

AllowedMethodsParser
Parse AllowedMethods from XML

AllowedMethodsWriter
Write AllowedMethods contents to a SignedRequest

AllowedOriginParser
Parse AllowedOrigin from XML

AllowedOriginWriter
Write AllowedOrigin contents to a SignedRequest

AllowedOriginsParser
Parse AllowedOrigins from XML

AllowedOriginsWriter
Write AllowedOrigins contents to a SignedRequest

BucketLoggingStatusParser
Parse BucketLoggingStatus from XML

BucketLoggingStatusWriter
Write BucketLoggingStatus contents to a SignedRequest

BucketLogsPermissionParser
Parse BucketLogsPermission from XML

BucketLogsPermissionWriter
Write BucketLogsPermission contents to a SignedRequest

BucketVersioningStatusParser
Parse BucketVersioningStatus from XML

BucketVersioningStatusWriter
Write BucketVersioningStatus contents to a SignedRequest

CORSRuleParser
Parse CORSRule from XML

CORSRuleWriter
Write CORSRule contents to a SignedRequest

CORSRulesParser
Parse CORSRules from XML

CORSRulesWriter
Write CORSRules contents to a SignedRequest

CacheControlParser
Parse CacheControl from XML

CacheControlWriter
Write CacheControl contents to a SignedRequest

CloudFunctionConfigurationParser
Parse CloudFunctionConfiguration from XML

CloudFunctionConfigurationWriter
Write CloudFunctionConfiguration contents to a SignedRequest

CloudFunctionInvocationRoleParser
Parse CloudFunctionInvocationRole from XML

CloudFunctionInvocationRoleWriter
Write CloudFunctionInvocationRole contents to a SignedRequest

CloudFunctionParser
Parse CloudFunction from XML

CloudFunctionWriter
Write CloudFunction contents to a SignedRequest

CommonPrefix

CommonPrefixListParser
Parse CommonPrefixList from XML

CommonPrefixListWriter
Write CommonPrefixList contents to a SignedRequest

CommonPrefixParser
Parse CommonPrefix from XML

CommonPrefixWriter
Write CommonPrefix contents to a SignedRequest

Condition

ConditionParser
Parse Condition from XML

ConditionWriter
Write Condition contents to a SignedRequest

ContentDispositionParser
Parse ContentDisposition from XML

ContentDispositionWriter
Write ContentDisposition contents to a SignedRequest

ContentEncodingParser
Parse ContentEncoding from XML

ContentEncodingWriter
Write ContentEncoding contents to a SignedRequest

ContentLanguageParser
Parse ContentLanguage from XML

ContentLanguageWriter
Write ContentLanguage contents to a SignedRequest

ContentLengthParser
Parse ContentLength from XML

ContentLengthWriter
Write ContentLength contents to a SignedRequest

CopySourceIfMatchParser
Parse CopySourceIfMatch from XML

CopySourceIfMatchWriter
Write CopySourceIfMatch contents to a SignedRequest

CopySourceIfModifiedSinceParser
Parse CopySourceIfModifiedSince from XML

CopySourceIfModifiedSinceWriter
Write CopySourceIfModifiedSince contents to a SignedRequest

CopySourceIfNoneMatchParser
Parse CopySourceIfNoneMatch from XML

CopySourceIfNoneMatchWriter
Write CopySourceIfNoneMatch contents to a SignedRequest

CopySourceIfUnmodifiedSinceParser
Parse CopySourceIfUnmodifiedSince from XML

CopySourceIfUnmodifiedSinceWriter
Write CopySourceIfUnmodifiedSince contents to a SignedRequest

CopySourceParser
Parse CopySource from XML

CopySourceSSECustomerAlgorithmParser
Parse CopySourceSSECustomerAlgorithm from XML

CopySourceSSECustomerAlgorithmWriter
Write CopySourceSSECustomerAlgorithm contents to a SignedRequest

CopySourceSSECustomerKeyMD5Parser
Parse CopySourceSSECustomerKeyMD5 from XML

CopySourceSSECustomerKeyMD5Writer
Write CopySourceSSECustomerKeyMD5 contents to a SignedRequest

CopySourceSSECustomerKeyParser
Parse CopySourceSSECustomerKey from XML

CopySourceSSECustomerKeyWriter
Write CopySourceSSECustomerKey contents to a SignedRequest

CopySourceVersionIdParser
Parse CopySourceVersionId from XML

CopySourceVersionIdWriter
Write CopySourceVersionId contents to a SignedRequest

CopySourceWriter
Write CopySource contents to a SignedRequest

CreateBucketConfiguration

CreateBucketOutput

CreateBucketOutputParser
Parse CreateBucketOutput from XML

CreateBucketOutputWriter
Write CreateBucketOutput contents to a SignedRequest

DateParser
Parse Date from XML

DateWriter
Write Date contents to a SignedRequest

DaysParser
Parse Days from XML

DaysWriter
Write Days contents to a SignedRequest

DeleteBucketCorsRequestParser
Parse DeleteBucketCorsRequest from XML

DeleteBucketCorsRequestWriter
Write DeleteBucketCorsRequest contents to a SignedRequest

DeleteBucketPolicyRequestParser
Parse DeleteBucketPolicyRequest from XML

DeleteBucketPolicyRequestWriter
Write DeleteBucketPolicyRequest contents to a SignedRequest

DeleteBucketReplicationRequestParser
Parse DeleteBucketReplicationRequest from XML

DeleteBucketReplicationRequestWriter
Write DeleteBucketReplicationRequest contents to a SignedRequest

DeleteBucketRequestParser
Parse DeleteBucketRequest from XML

DeleteBucketTaggingRequestParser
Parse DeleteBucketTaggingRequest from XML

DeleteBucketTaggingRequestWriter
Write DeleteBucketTaggingRequest contents to a SignedRequest

DeleteBucketWebsiteRequestParser
Parse DeleteBucketWebsiteRequest from XML

DeleteBucketWebsiteRequestWriter
Write DeleteBucketWebsiteRequest contents to a SignedRequest

DeleteMarkerParser
Parse DeleteMarker from XML

DeleteMarkerWriter
Write DeleteMarker contents to a SignedRequest

DelimiterParser
Parse Delimiter from XML

DelimiterWriter
Write Delimiter contents to a SignedRequest

DestinationParser
Parse Destination from XML

DestinationWriter
Write Destination contents to a SignedRequest

EncodingTypeParser
Parse EncodingType from XML

EncodingTypeWriter
Write EncodingType contents to a SignedRequest

ErrorDocumentParser
Parse ErrorDocument from XML

ErrorDocumentWriter
Write ErrorDocument contents to a SignedRequest

ErrorsParser
Parse Errors from XML

ErrorsWriter
Write Errors contents to a SignedRequest

EventListParser
Parse EventList from XML

EventListWriter
Write EventList contents to a SignedRequest

EventParser
Parse Event from XML

EventWriter
Write Event contents to a SignedRequest

ExpirationParser
Parse Expiration from XML

ExpirationStatusParser
Parse ExpirationStatus from XML

ExpirationStatusWriter
Write ExpirationStatus contents to a SignedRequest

ExpirationWriter
Write Expiration contents to a SignedRequest

ExposeHeaderParser
Parse ExposeHeader from XML

ExposeHeaderWriter
Write ExposeHeader contents to a SignedRequest

ExposeHeadersParser
Parse ExposeHeaders from XML

ExposeHeadersWriter
Write ExposeHeaders contents to a SignedRequest

GetBucketCorsOutputParser
Parse GetBucketCorsOutput from XML

GetBucketCorsOutputWriter
Write GetBucketCorsOutput contents to a SignedRequest

GetBucketCorsRequestParser
Parse GetBucketCorsRequest from XML

GetBucketCorsRequestWriter
Write GetBucketCorsRequest contents to a SignedRequest

GetBucketLoggingOutputParser
Parse GetBucketLoggingOutput from XML

GetBucketLoggingOutputWriter
Write GetBucketLoggingOutput contents to a SignedRequest

GetBucketLoggingRequestParser
Parse GetBucketLoggingRequest from XML

GetBucketLoggingRequestWriter
Write GetBucketLoggingRequest contents to a SignedRequest

GetBucketNotificationConfigurationRequestParser
Parse GetBucketNotificationConfigurationRequest from XML

GetBucketNotificationConfigurationRequestWriter
Write GetBucketNotificationConfigurationRequest contents to a SignedRequest

GetBucketReplicationOutputParser
Parse GetBucketReplicationOutput from XML

GetBucketReplicationOutputWriter
Write GetBucketReplicationOutput contents to a SignedRequest

GetBucketVersioningOutputParser
Parse GetBucketVersioningOutput from XML

GetBucketVersioningOutputWriter
Write GetBucketVersioningOutput contents to a SignedRequest

GetBucketVersioningRequestParser
Parse GetBucketVersioningRequest from XML

GetBucketVersioningRequestWriter
Write GetBucketVersioningRequest contents to a SignedRequest

GetBucketWebsiteRequest

GetBucketWebsiteRequestParser
Parse GetBucketWebsiteRequest from XML

GetBucketWebsiteRequestWriter
Write GetBucketWebsiteRequest contents to a SignedRequest

IfMatchParser
Parse IfMatch from XML

IfMatchWriter
Write IfMatch contents to a SignedRequest

IfModifiedSinceParser
Parse IfModifiedSince from XML

IfModifiedSinceWriter
Write IfModifiedSince contents to a SignedRequest

IfNoneMatchParser
Parse IfNoneMatch from XML

IfNoneMatchWriter
Write IfNoneMatch contents to a SignedRequest

IfUnmodifiedSinceParser
Parse IfUnmodifiedSince from XML

IfUnmodifiedSinceWriter
Write IfUnmodifiedSince contents to a SignedRequest

IndexDocumentParser
Parse IndexDocument from XML

IndexDocumentWriter
Write IndexDocument contents to a SignedRequest

InitiatedParser
Parse Initiated from XML

InitiatedWriter
Write Initiated contents to a SignedRequest

InitiatorParser
Parse Initiator from XML

InitiatorWriter
Write Initiator contents to a SignedRequest

IsLatestParser
Parse IsLatest from XML

IsLatestWriter
Write IsLatest contents to a SignedRequest

KeyMarkerParser
Parse KeyMarker from XML

KeyMarkerWriter
Write KeyMarker contents to a SignedRequest

KeyPrefixEqualsParser
Parse KeyPrefixEquals from XML

KeyPrefixEqualsWriter
Write KeyPrefixEquals contents to a SignedRequest

LambdaFunctionArnParser
Parse LambdaFunctionArn from XML

LambdaFunctionArnWriter
Write LambdaFunctionArn contents to a SignedRequest

LambdaFunctionConfigurationListParser
Parse LambdaFunctionConfigurationList from XML

LambdaFunctionConfigurationListWriter
Write LambdaFunctionConfigurationList contents to a SignedRequest

LambdaFunctionConfigurationParser
Parse LambdaFunctionConfiguration from XML

LambdaFunctionConfigurationWriter
Write LambdaFunctionConfiguration contents to a SignedRequest

LastModifiedParser
Parse LastModified from XML

LastModifiedWriter
Write LastModified contents to a SignedRequest

LifecycleConfigurationParser
Parse LifecycleConfiguration from XML

LifecycleConfigurationWriter
Write LifecycleConfiguration contents to a SignedRequest

LifecycleExpiration

LifecycleExpirationParser
Parse LifecycleExpiration from XML

LifecycleExpirationWriter
Write LifecycleExpiration contents to a SignedRequest

ListBucketsOutput

ListBucketsOutputParser
Parse ListBucketsOutput from XML

ListBucketsOutputWriter
Write ListBucketsOutput contents to a SignedRequest

LocationParser
Parse Location from XML

LocationWriter
Write Location contents to a SignedRequest

MFADeleteParser
Parse MFADelete from XML

MFADeleteStatusParser
Parse MFADeleteStatus from XML

MFADeleteStatusWriter
Write MFADeleteStatus contents to a SignedRequest

MFADeleteWriter
Write MFADelete contents to a SignedRequest

MFAParser
Parse MFA from XML

MFAWriter
Write MFA contents to a SignedRequest

MarkerParser
Parse Marker from XML

MarkerWriter
Write Marker contents to a SignedRequest

MaxAgeSecondsParser
Parse MaxAgeSeconds from XML

MaxAgeSecondsWriter
Write MaxAgeSeconds contents to a SignedRequest

MetadataDirectiveParser
Parse MetadataDirective from XML

MetadataDirectiveWriter
Write MetadataDirective contents to a SignedRequest

MetadataKeyParser
Parse MetadataKey from XML

MetadataKeyWriter
Write MetadataKey contents to a SignedRequest

MetadataParser
Parse Metadata from XML

MetadataValueParser
Parse MetadataValue from XML

MetadataValueWriter
Write MetadataValue contents to a SignedRequest

MetadataWriter
Write Metadata contents to a SignedRequest

MissingMetaParser
Parse MissingMeta from XML

MissingMetaWriter
Write MissingMeta contents to a SignedRequest

MultipartUploadIdParser
Parse MultipartUploadId from XML

MultipartUploadIdWriter
Write MultipartUploadId contents to a SignedRequest

NextKeyMarkerParser
Parse NextKeyMarker from XML

NextKeyMarkerWriter
Write NextKeyMarker contents to a SignedRequest

NextMarkerParser
Parse NextMarker from XML

NextMarkerWriter
Write NextMarker contents to a SignedRequest

NoncurrentVersionExpirationParser
Parse NoncurrentVersionExpiration from XML

NoncurrentVersionExpirationWriter
Write NoncurrentVersionExpiration contents to a SignedRequest

NoncurrentVersionTransition
Container for the transition rule that describes when noncurrent objects transition to the GLACIER storage class. If your bucket is versioning-enabled (or versioning is suspended), you can set this action to request that Amazon S3 transition noncurrent object versions to the GLACIER storage class at a specific period in the object's lifetime.

NoncurrentVersionTransitionParser
Parse NoncurrentVersionTransition from XML

NoncurrentVersionTransitionWriter
Write NoncurrentVersionTransition contents to a SignedRequest

NotificationConfigurationDeprecated

NotificationConfigurationDeprecatedParser
Parse NotificationConfigurationDeprecated from XML

NotificationConfigurationDeprecatedWriter
Write NotificationConfigurationDeprecated contents to a SignedRequest

NotificationIdParser
Parse NotificationId from XML

NotificationIdWriter
Write NotificationId contents to a SignedRequest

PayerParser
Parse Payer from XML

PayerWriter
Write Payer contents to a SignedRequest

PolicyParser
Parse Policy from XML

PolicyWriter
Write Policy contents to a SignedRequest

PutBucketLifecycleRequestParser
Parse PutBucketLifecycleRequest from XML

PutBucketLifecycleRequestWriter
Write PutBucketLifecycleRequest contents to a SignedRequest

PutBucketLoggingRequestParser
Parse PutBucketLoggingRequest from XML

PutBucketLoggingRequestWriter
Write PutBucketLoggingRequest contents to a SignedRequest

PutBucketNotificationRequestParser
Parse PutBucketNotificationRequest from XML

PutBucketNotificationRequestWriter
Write PutBucketNotificationRequest contents to a SignedRequest

PutBucketPolicyRequestParser
Parse PutBucketPolicyRequest from XML

PutBucketPolicyRequestWriter
Write PutBucketPolicyRequest contents to a SignedRequest

PutBucketReplicationRequestParser
Parse PutBucketReplicationRequest from XML

PutBucketReplicationRequestWriter
Write PutBucketReplicationRequest contents to a SignedRequest

PutBucketRequestPaymentRequestParser
Parse PutBucketRequestPaymentRequest from XML

PutBucketRequestPaymentRequestWriter
Write PutBucketRequestPaymentRequest contents to a SignedRequest

PutBucketVersioningRequestParser
Parse PutBucketVersioningRequest from XML

PutBucketVersioningRequestWriter
Write PutBucketVersioningRequest contents to a SignedRequest

PutBucketWebsiteRequestParser
Parse PutBucketWebsiteRequest from XML

PutBucketWebsiteRequestWriter
Write PutBucketWebsiteRequest contents to a SignedRequest

QueueArnParser
Parse QueueArn from XML

QueueArnWriter
Write QueueArn contents to a SignedRequest

QueueConfigurationDeprecatedParser
Parse QueueConfigurationDeprecated from XML

QueueConfigurationDeprecatedWriter
Write QueueConfigurationDeprecated contents to a SignedRequest

QueueConfigurationListParser
Parse QueueConfigurationList from XML

QueueConfigurationListWriter
Write QueueConfigurationList contents to a SignedRequest

QueueConfigurationParser
Parse QueueConfiguration from XML

QueueConfigurationWriter
Write QueueConfiguration contents to a SignedRequest

RedirectAllRequestsToParser
Parse RedirectAllRequestsTo from XML

RedirectAllRequestsToWriter
Write RedirectAllRequestsTo contents to a SignedRequest

RedirectParser
Parse Redirect from XML

RedirectWriter
Write Redirect contents to a SignedRequest

ReplaceKeyPrefixWithParser
Parse ReplaceKeyPrefixWith from XML

ReplaceKeyPrefixWithWriter
Write ReplaceKeyPrefixWith contents to a SignedRequest

ReplaceKeyWithParser
Parse ReplaceKeyWith from XML

ReplaceKeyWithWriter
Write ReplaceKeyWith contents to a SignedRequest

ReplicationConfigurationParser
Parse ReplicationConfiguration from XML

ReplicationConfigurationWriter
Write ReplicationConfiguration contents to a SignedRequest

ReplicationRuleParser
Parse ReplicationRule from XML

ReplicationRuleStatusParser
Parse ReplicationRuleStatus from XML

ReplicationRuleStatusWriter
Write ReplicationRuleStatus contents to a SignedRequest

ReplicationRuleWriter
Write ReplicationRule contents to a SignedRequest

ReplicationRulesParser
Parse ReplicationRules from XML

ReplicationRulesWriter
Write ReplicationRules contents to a SignedRequest

ReplicationStatusParser
Parse ReplicationStatus from XML

ReplicationStatusWriter
Write ReplicationStatus contents to a SignedRequest

RequestChargedParser
Parse RequestCharged from XML

RequestChargedWriter
Write RequestCharged contents to a SignedRequest

RequestPayerParser
Parse RequestPayer from XML

RequestPayerWriter
Write RequestPayer contents to a SignedRequest

RequestPaymentConfiguration

RequestPaymentConfigurationParser
Parse RequestPaymentConfiguration from XML

RequestPaymentConfigurationWriter
Write RequestPaymentConfiguration contents to a SignedRequest

ResponseCacheControlParser
Parse ResponseCacheControl from XML

ResponseCacheControlWriter
Write ResponseCacheControl contents to a SignedRequest

ResponseContentDispositionParser
Parse ResponseContentDisposition from XML

ResponseContentDispositionWriter
Write ResponseContentDisposition contents to a SignedRequest

ResponseContentEncodingParser
Parse ResponseContentEncoding from XML

ResponseContentEncodingWriter
Write ResponseContentEncoding contents to a SignedRequest

ResponseContentLanguageParser
Parse ResponseContentLanguage from XML

ResponseContentLanguageWriter
Write ResponseContentLanguage contents to a SignedRequest

ResponseContentTypeParser
Parse ResponseContentType from XML

ResponseContentTypeWriter
Write ResponseContentType contents to a SignedRequest

ResponseExpiresParser
Parse ResponseExpires from XML

ResponseExpiresWriter
Write ResponseExpires contents to a SignedRequest

RestoreParser
Parse Restore from XML

RestoreWriter
Write Restore contents to a SignedRequest

RoleParser
Parse Role from XML

RoleWriter
Write Role contents to a SignedRequest

RoutingRuleParser
Parse RoutingRule from XML

RoutingRuleWriter
Write RoutingRule contents to a SignedRequest

RoutingRulesParser
Parse RoutingRules from XML

RoutingRulesWriter
Write RoutingRules contents to a SignedRequest

Rule

RuleWriter
Write Rule contents to a SignedRequest

RulesParser
Parse Rules from XML

RulesWriter
Write Rules contents to a SignedRequest

S3ClientMessageParser
Parse S3ClientMessage from XML

S3ClientMessageWriter
Write S3ClientMessage contents to a SignedRequest

SSECustomerAlgorithmParser
Parse SSECustomerAlgorithm from XML

SSECustomerAlgorithmWriter
Write SSECustomerAlgorithm contents to a SignedRequest

SSECustomerKeyMD5Parser
Parse SSECustomerKeyMD5 from XML

SSECustomerKeyMD5Writer
Write SSECustomerKeyMD5 contents to a SignedRequest

SSECustomerKeyParser
Parse SSECustomerKey from XML

SSECustomerKeyWriter
Write SSECustomerKey contents to a SignedRequest

SSEKMSKeyIdParser
Parse SSEKMSKeyId from XML

SSEKMSKeyIdWriter
Write SSEKMSKeyId contents to a SignedRequest

ServerSideEncryptionParser
Parse ServerSideEncryption from XML

ServerSideEncryptionWriter
Write ServerSideEncryption contents to a SignedRequest

SizeParser
Parse Size from XML

SizeWriter
Write Size contents to a SignedRequest

StorageClassParser
Parse StorageClass from XML

StorageClassWriter
Write StorageClass contents to a SignedRequest

SuffixParser
Parse Suffix from XML

SuffixWriter
Write Suffix contents to a SignedRequest

TopicArnParser
Parse TopicArn from XML

TopicArnWriter
Write TopicArn contents to a SignedRequest

TopicConfigurationDeprecatedParser
Parse TopicConfigurationDeprecated from XML

TopicConfigurationDeprecatedWriter
Write TopicConfigurationDeprecated contents to a SignedRequest

Transition

TransitionParser
Parse Transition from XML

TransitionStorageClassParser
Parse TransitionStorageClass from XML

TransitionStorageClassWriter
Write TransitionStorageClass contents to a SignedRequest

TransitionWriter
Write Transition contents to a SignedRequest

VersionIdMarkerParser
Parse VersionIdMarker from XML

VersionIdMarkerWriter
Write VersionIdMarker contents to a SignedRequest

VersioningConfigurationParser
Parse VersioningConfiguration from XML

VersioningConfigurationWriter
Write VersioningConfiguration contents to a SignedRequest

WebsiteConfigurationParser
Parse WebsiteConfiguration from XML

WebsiteConfigurationWriter
Write WebsiteConfiguration contents to a SignedRequest

WebsiteRedirectLocationParser
Parse WebsiteRedirectLocation from XML

WebsiteRedirectLocationWriter
Write WebsiteRedirectLocation contents to a SignedRequest

Type Definitions

AllowedMethod

AllowedMethods

AllowedOrigin

AllowedOrigins

BucketLogsPermission

BucketVersioningStatus

CORSRules

CacheControl

CloudFunction

CloudFunctionInvocationRole

CommonPrefixList
Confirms that the requester knows that she or he will be charged for the request. Bucket owners need not specify this parameter in their requests. Documentation on downloading objects from requester pays buckets can be found here.

ContentDisposition

ContentEncoding

ContentLanguage

ContentLength

CopySource

CopySourceIfMatch

CopySourceIfModifiedSince

CopySourceIfNoneMatch

CopySourceIfUnmodifiedSince

CopySourceSSECustomerAlgorithm

CopySourceSSECustomerKey

CopySourceSSECustomerKeyMD5

CopySourceVersionId

Date

Days

DeleteMarker

Delimiter

EncodingType
Requests Amazon S3 to encode the object keys in the response and specifies the encoding method to use. An object key may contain any Unicode character; however, XML 1.0 parser cannot parse some characters, such as characters with an ASCII value from 0 to 10. For characters that are not supported in XML 1.0, you can add this parameter to request that Amazon S3 encode the keys in the response.

Errors

Event
Bucket event for which to send notifications.

EventList

Expiration

ExpirationStatus

ExposeHeader

ExposeHeaders

IfMatch

IfModifiedSince

IfNoneMatch

IfUnmodifiedSince

Initiated

IsLatest

KeyMarker

KeyPrefixEquals

LambdaFunctionArn

LambdaFunctionConfigurationList

LastModified

Location

MFA

MFADelete

MFADeleteStatus

Marker

MaxAgeSeconds

Metadata

MetadataDirective

MetadataKey

MetadataValue

MissingMeta

MultipartUploadId

NextKeyMarker

NextMarker

NotificationId
Optional unique identifier for configurations in a notification configuration. If you don't provide one, Amazon S3 will assign an ID.

Payer

QueueArn

QueueConfigurationList

ReplaceKeyPrefixWith

ReplaceKeyWith

ReplicationRuleStatus

ReplicationRules

ReplicationStatus

RequestCharged
If present, indicates that the requester was successfully charged for the request.

RequestPayer
Confirms that the requester knows that she or he will be charged for the request. Bucket owners need not specify this parameter in their requests. Documentation on downloading objects from requester pays buckets can be found here.

ResponseCacheControl

ResponseContentDisposition

ResponseContentEncoding

ResponseContentLanguage

ResponseContentType

ResponseExpires

Restore

Role

RoutingRules

Rules

S3ClientMessage

SSECustomerAlgorithm

SSECustomerKey

SSECustomerKeyMD5

SSEKMSKeyId

ServerSideEncryption

Size

StorageClass

Suffix

TopicArn

TransitionStorageClass

VersionIdMarker

WebsiteRedirectLocation