Struct aws_sdk_rust::aws::s3::writeparse::Condition

pub struct Condition {
    pub http_error_code_returned_equals: HttpErrorCodeReturnedEquals,
    pub key_prefix_equals: KeyPrefixEquals,
}

Fields
http_error_code_returned_equals: HttpErrorCodeReturnedEquals

The HTTP error code when the redirect is applied. In the event of an error, if the error code equals this value, then the specified redirect is applied. Required when parent element Condition is specified and sibling KeyPrefixEquals is not specified. If both are specified, then both must be true for the redirect to be applied.

key_prefix_equals: KeyPrefixEquals

The object key name prefix when the redirect is applied. For example, to redirect requests for ExamplePage.html, the key prefix will be ExamplePage.html. To redirect request for all pages with the prefix docs/, the key prefix will be /docs, which identifies all objects in the docs/ folder. Required when the parent element Condition is specified and sibling HttpErrorCodeReturnedEquals is not specified. If both conditions are specified, both must be true for the redirect to be applied.