Struct aws_sdk_rust::aws::s3::writeparse::NoncurrentVersionTransition

pub struct NoncurrentVersionTransition {
    pub noncurrent_days: Days,
    pub storage_class: TransitionStorageClass,
}

Container for the transition rule that describes when noncurrent objects transition to the GLACIER storage class. If your bucket is versioning-enabled (or versioning is suspended), you can set this action to request that Amazon S3 transition noncurrent object versions to the GLACIER storage class at a specific period in the object's lifetime.

Fields
noncurrent_days: Days

Specifies the number of days an object is noncurrent before Amazon S3 can perform the associated action. For information about the noncurrent days calculations, see How Amazon S3 Calculates When an Object Became Noncurrent in the Amazon Simple Storage Service Developer Guide.

storage_class: TransitionStorageClass

The class of storage used to store the object.