Struct aws_sdk_rust::aws::s3::object::ListObjectsOutput

pub struct ListObjectsOutput {
    pub name: BucketName,
    pub next_marker: NextMarker,
    pub delimiter: Delimiter,
    pub max_keys: MaxKeys,
    pub prefix: Prefix,
    pub marker: Marker,
    pub encoding_type: EncodingType,
    pub is_truncated: IsTruncated,
    pub contents: ObjectMetadataList,
    pub common_prefixes: CommonPrefixList,
    pub continuation_token: ContinuationToken,
    pub next_continuation_token: ContinuationToken,
    pub key_count: KeyCount,
    pub start_after: StartAfter,
}

ListObjectsOutput contains the list of objects and their associated metadata for a given bucket name. There are two versions, version 1 and version 2. AWS S3 supports both. The struct field names that are unique for Version 1 are marked and those that are unique for Version 2 are marked. Those not marked are common between versions. The default for AWS is version 2 but you have to specify version 2 in ListObjectsRequest or it will default to version 1.

Fields
name: BucketName next_marker: NextMarker

Version 1. When response is truncated (the IsTruncated element value in the response is true), you can use the key name in this field as marker in the subsequent request to get next set of objects. Amazon S3 lists objects in alphabetical order Note: This element is returned only if you have delimiter request parameter specified. If response does not include the NextMaker and it is truncated, you can use the value of the last Key in the response as the marker in the subsequent request to get the next set of object keys.

delimiter: Delimiter max_keys: MaxKeys prefix: Prefix marker: Marker

Version 1.

encoding_type: EncodingType

Encoding type used by Amazon S3 to encode object keys in the response.

is_truncated: IsTruncated

A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria.

contents: ObjectMetadataList

List of ObjectMetadata for each object in the given bucket.

common_prefixes: CommonPrefixList continuation_token: ContinuationToken

Version 2. Returned if included in the request

next_continuation_token: ContinuationToken

Version 2. If the response is truncated, Amazon S3 returns this parameter with a continuation token that you can specify as the continuation-token in your next request to retrieve the next set of keys.

key_count: KeyCount

Version 2. Returned number of keys in response. Always <= MaxKeys.

start_after: StartAfter

Version 2. Is included with the response if sent with the request.